Will move 'LocalFeedLoader' (and 'FeedStore' collaborator) to its own file in production
